Add build-flag note for mbed.
The Arduino tutorial has a section which lets you know you have to set the build flags, but there is no such thing for mbed. Being a noob, I was trying to find why even the copy/pasted examples were failing in mbed. And found out I have to add the flag when looking at the example's .ini file.

I think adding this note would be helpful to people getting started. Please feel free to modify the commit if you feel it could be done better.