Add "serial" upload information
Serial uploading using the internal STM bootloader is available in 4.1.0 when upload_protocol is set to "serial" and an USB/TTL-UART adapter is connected to pins PA9 (TX) and PA10 (RX) and the Blue Pill is reset with jumpers BOOT0=1 and BOOT1=0. I think the docs should reflect that as it is a very common and easy method.