Update lib_deps.rst
specifying "can be left blank", just as a point of reference. I left it blank on a copy+paste platformio.ini template I made, and it successfully compiles. I had to test it to see - it'd be nice to know beforehand if blank is okay for compilation, adding this for the next guy :)